+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - fixed pickup point map initialization error on checkout when carrier is selected by default and JS files are loaded deferred
+- fixed validation error message display at the top of the page before appearing in the carrier section on checkout
 
 ## [1.2.1] - 2026-05-05
 ### Fixed